Add button to add dataset to collection in collection view
