+++ src/layout/BreadCrumbs.vue
@@ -27,13 +27,13 @@
                 v-if="item.to.name == 'dataset' && configurationId"
               >
                 <v-divider class="my-1" />
-                <a
-                  class="d-flex align-center justify-center body-2"
+                <div
+                  class="clickable-flex justify-center body-2"
                   @click="openAddDatasetDialog(configurationId)"
                 >
-                  <v-icon>mdi-plus</v-icon>
+                  <v-icon class="pr-2" color="primary">mdi-plus-circle</v-icon>
                   Add a dataset in this collection
-                </a>
+                </div>
               </template>
             </v-select>
             <router-link :to="item.to">
@@ -53,109 +53,12 @@
     </v-breadcrumbs>
 
     <!-- Dialog to add a dataset to the current collection -->
-    <v-dialog v-model="addDatasetDialog" v-if="addDatasetConfig" width="60%">
-      <v-card>
-        <v-card-title class="mb-4">
-          <span class="text--secondary">Adding dataset to collection:</span>
-          <span class="text--primary">{{ addDatasetConfig.name }}</span>
-        </v-card-title>
-        <v-card-text>
-          <v-radio-group v-model="addDatasetOptionType">
-            <v-radio
-              label="Add an existing dataset to the current collection"
-              value="add"
-            />
-            <v-radio
-              label="Upload new dataset and add to current collection"
-              value="upload"
-            />
-          </v-radio-group>
-          <template v-if="addDatasetOption.type == 'add'">
-            <custom-file-manager
-              title="Choose a dataset or a folder of datasets to add in the current Configuration"
-              :breadcrumb="true"
-              :selectable="true"
-              @selected="selectAddDatasetFolder"
-              :location.sync="datasetSelectLocation"
-              :initial-items-per-page="-1"
-              :items-per-page-options="[-1]"
-              :menu-enabled="false"
-              :more-chips="false"
-              :clickable-chips="false"
-            />
-            <v-alert
-              v-for="(warning, iWarning) in addDatasetOption.warnings"
-              :key="iWarning + '-warning'"
-              type="warning"
-              class="my-2"
-              dense
-            >
-              {{ warning }}
-            </v-alert>
-            <v-alert
-              v-if="addDatasetOption.datasets.length > 0"
-              type="success"
-              class="my-2"
-              dense
-            >
-              Selected {{ addDatasetOption.datasets.length }} dataset(s):
-              <v-divider />
-              <div
-                v-for="dataset in addDatasetOption.datasets"
-                :key="'selected-' + dataset.id"
-              >
-                {{ dataset.name }}
-                <v-divider />
-              </div>
-            </v-alert>
-            <v-card-actions class="ma-2">
-              <v-spacer />
-              <v-btn
-                color="green"
-                :disabled="!canAddDatasetToCollection"
-                @click="addDatasetToCollection"
-              >
-                Add dataset
-              </v-btn>
-            </v-card-actions>
-          </template>
-          <template v-if="addDatasetOption.type == 'upload'">
-            <new-dataset
-              @datasetUploaded="addDatasetToCollectionUploaded"
-              :autoMultiConfig="false"
-            />
-            <div class="d-flex justify-end mt-2">
-              <v-simple-checkbox
-                v-model="addDatasetOption.editVariables"
-                dense
-              />
-              Review variables
-            </div>
-            <tempalte v-if="addDatasetOption.configuring">
-              <multi-source-configuration
-                ref="configuration"
-                :datasetId="addDatasetOption.uploadedDatasetId"
-                :autoDatasetRoute="false"
-                @log="addDatasetOption.configurationLogs = $event"
-                @generatedJson="addDatasetConfigurationDone"
-                :class="{ 'd-none': !addDatasetOption.editVariables }"
-              />
-              <div class="title mb-2">
-                Configuring the dataset
-              </div>
-              <v-progress-circular indeterminate />
-              <div
-                class="code-container"
-                v-if="addDatasetOption.configurationLogs"
-              >
-                <code class="code-block">
-                  {{ addDatasetOption.configurationLogs }}
-                </code>
-              </div>
-            </tempalte>
-          </template>
-        </v-card-text>
-      </v-card>
+    <v-dialog v-model="addDatasetDialog" width="60%">
+      <add-dataset-to-collection
+        v-if="addDatasetCollection"
+        :collection="addDatasetCollection"
+        @addedDatasets="addedDatasets"
+      />
     </v-dialog>
   </div>
 </template>
@@ -166,15 +69,9 @@ import store from "@/store";
 import girderResources from "@/store/girderResources";
 import { Location } from "vue-router";
 import { Dictionary } from "vue-router/types/router";
-import { IDataset, IDatasetConfiguration, areCompatibles } from "@/store/model";
-import CustomFileManager from "@/components/CustomFileManager.vue";
+import AddDatasetToCollection from "@/components/AddDatasetToCollection.vue";
 
-import NewDataset from "@/views/dataset/NewDataset.vue";
-import { IGirderSelectAble } from "@/girder";
-import { isDatasetFolder } from "@/utils/girderSelectable";
-import { getDatasetCompatibility } from "@/store/GirderAPI";
-import { logError } from "@/utils/log";
-import MultiSourceConfiguration from "@/views/dataset/MultiSourceConfiguration.vue";
+import { IDatasetConfiguration } from "@/store/model";
 
 interface IBreadCrumbItem {
   title: string;
@@ -186,41 +83,7 @@ interface IBreadCrumbItem {
   }[];
 }
 
-type TAddDatasetOption =
-  | {
-      type: "upload";
-      editVariables: boolean;
-      configuring: boolean;
-      configurationLogs: string;
-      uploadedDatasetId: string | null;
-    }
-  | {
-      type: "add";
-      datasets: IDataset[];
-      warnings: string[];
-    };
-
-function defaultDatasetUploadOption(): TAddDatasetOption {
-  return {
-    type: "upload",
-    editVariables: false,
-    configuring: false,
-    configurationLogs: "",
-    uploadedDatasetId: null
-  };
-}
-
-function defaultDatasetAddOption(): TAddDatasetOption {
-  return {
-    type: "add",
-    datasets: [],
-    warnings: []
-  };
-}
-
-@Component({
-  components: { CustomFileManager, MultiSourceConfiguration, NewDataset }
-})
+@Component({ components: { AddDatasetToCollection } })
 export default class BreadCrumbs extends Vue {
   readonly store = store;
   readonly girderResources = girderResources;
@@ -236,220 +99,7 @@ export default class BreadCrumbs extends Vue {
     routeName: null
   };
 
-  addDatasetConfig: IDatasetConfiguration | null = null;
-  addDatasetOption: TAddDatasetOption = defaultDatasetUploadOption();
-
-  // Enable navigation in the component
-  datasetSelectLocation: IGirderSelectAble | null = null;
-
-  $refs!: {
-    configuration?: MultiSourceConfiguration;
-  };
-
-  get addDatasetOptionType(): TAddDatasetOption["type"] {
-    return this.addDatasetOption.type;
-  }
-
-  set addDatasetOptionType(type: TAddDatasetOption["type"]) {
-    switch (type) {
-      case "add":
-        this.addDatasetOption = defaultDatasetAddOption();
-        break;
-      case "upload":
-        this.addDatasetOption = defaultDatasetUploadOption();
-        break;
-    }
-  }
-
-  async openAddDatasetDialog(configIdPromise: Promise<string>) {
-    const configId = await configIdPromise;
-    this.addDatasetConfig = await this.girderResources.getConfiguration(
-      configId
-    );
-    this.addDatasetOption = defaultDatasetUploadOption();
-  }
-
-  get addDatasetDialog() {
-    return this.addDatasetConfig !== null;
-  }
-
-  set addDatasetDialog(value: boolean) {
-    if (!value) {
-      this.addDatasetConfig = null;
-    }
-  }
-
-  async selectAddDatasetFolder(selectedLocations: IGirderSelectAble[]) {
-    const option = this.addDatasetOption;
-    if (option.type !== "add") {
-      return;
-    }
-    if (selectedLocations.length === 0 || !this.addDatasetConfig) {
-      option.datasets = [];
-      option.warnings = [];
-      return;
-    }
-
-    const currentWarnings: string[] = [];
-
-    // Find selected items that are datasets
-    const selectedDatasets: IDataset[] = [];
-    await Promise.all(
-      selectedLocations.map(async location => {
-        if (!isDatasetFolder(location)) {
-          return;
-        }
-        const dataset = await this.girderResources.getDataset({
-          id: location._id
-        });
-        if (!dataset) {
-          return;
-        }
-        selectedDatasets.push(dataset);
-      })
-    );
-
-    if (selectedDatasets.length !== selectedLocations.length) {
-      const nNotDataset = selectedLocations.length - selectedDatasets.length;
-      currentWarnings.push(nNotDataset + " selected items are not datasests");
-    }
-
-    // Early return if no selected dataset
-    if (selectedDatasets.length === 0) {
-      option.datasets = [];
-      option.warnings = currentWarnings;
-      return;
-    }
-
-    // Filter compatible datasets
-    const configCompat = this.addDatasetConfig.compatibility;
-    const configViews = await this.store.api.findDatasetViews({
-      configurationId: this.addDatasetConfig.id
-    });
-    const excludeDatasetIds = new Set(configViews.map(view => view.datasetId));
-    const compatibleDatasets = selectedDatasets.filter(
-      dataset =>
-        !excludeDatasetIds.has(dataset.id) &&
-        areCompatibles(configCompat, getDatasetCompatibility(dataset))
-    );
-
-    if (compatibleDatasets.length !== selectedDatasets.length) {
-      const nNotCompatible =
-        selectedDatasets.length - compatibleDatasets.length;
-      currentWarnings.push(
-        nNotCompatible +
-          " selected items are not compatible with the current configuration"
-      );
-    }
-
-    // Return the dataset and warnings
-    option.datasets = compatibleDatasets;
-    option.warnings = currentWarnings;
-  }
-
-  get canAddDatasetToCollection(): boolean {
-    if (this.addDatasetConfig === null) {
-      return false;
-    }
-    switch (this.addDatasetOption.type) {
-      case "add":
-        return this.addDatasetOption.datasets.length > 0;
-      case "upload":
-        return false;
-    }
-  }
-
-  async addDatasetToCollectionUploaded(datasetId: string) {
-    if (this.addDatasetOption.type !== "upload") {
-      return;
-    }
-    this.addDatasetOption.uploadedDatasetId = datasetId;
-    this.addDatasetOption.configurationLogs = "";
-    this.addDatasetOption.configuring = true;
-
-    // Automatic generation of the JSON
-    if (!this.addDatasetOption.editVariables) {
-      await Vue.nextTick();
-      const config = this.$refs.configuration;
-      if (!config) {
-        logError(
-          "MultiSourceConfiguration component not mounted during configuration of new dataset"
-        );
-        this.addDatasetOption.editVariables = true;
-        return;
-      }
-      // Ensure that the component is initialized
-      await (config.initialized || config.initialize());
-      await config.submit();
-    }
-  }
-
-  async addDatasetConfigurationDone(jsonId: string | null) {
-    if (
-      this.addDatasetOption.type !== "upload" ||
-      !this.addDatasetOption.uploadedDatasetId
-    ) {
-      return;
-    }
-    this.addDatasetOption.configuring = false;
-    const datasetId = this.addDatasetOption.uploadedDatasetId;
-    if (!jsonId) {
-      logError("Failed to generate JSON during configuration of new dataset");
-      this.addDatasetOption.editVariables = true;
-      return;
-    }
-    // Check if dataset is compatible with configuration
-    const configCompat = this.addDatasetConfig?.compatibility;
-    if (!configCompat) {
-      logError(
-        "DatasetConfiguration missing after multi source configuration of dataset"
-      );
-      return;
-    }
-    const dataset = await this.girderResources.getDataset({ id: datasetId });
-    if (!dataset) {
-      logError("Dataset missing after multi source configuration of dataset");
-      return;
-    }
-    const datasetCompat = getDatasetCompatibility(dataset);
-    if (!areCompatibles(configCompat, datasetCompat)) {
-      logError("Incompatible dataset uploaded");
-      return;
-    }
-    // Set the dataset now that multi-source is available
-    await this.store.setSelectedDataset(datasetId);
-    this.addDatasetToCollection();
-  }
-
-  async addDatasetToCollection() {
-    if (!this.addDatasetConfig) {
-      return;
-    }
-    const configurationId = this.addDatasetConfig.id;
-    const datasetIds: string[] = [];
-
-    switch (this.addDatasetOption.type) {
-      case "add":
-        this.addDatasetOption.datasets.forEach(d => datasetIds.push(d.id));
-        break;
-      case "upload":
-        datasetIds.push(this.addDatasetOption.uploadedDatasetId!);
-        break;
-    }
-
-    const promises = datasetIds.map(datasetId =>
-      this.store.createDatasetView({
-        datasetId,
-        configurationId
-      })
-    );
-    await Promise.all(promises);
-
-    this.refreshItems(true);
-
-    this.addDatasetDialog = false;
-    this.addDatasetOption = defaultDatasetUploadOption();
-  }
+  addDatasetCollection: IDatasetConfiguration | null = null;
 
   get datasetView() {
     const { datasetViewId } = this.$route.params;
@@ -502,6 +152,29 @@ export default class BreadCrumbs extends Vue {
     if (resource) {
       Vue.set(item, "text", resource.name);
     }
+  }
+
+  async openAddDatasetDialog(configIdPromise: Promise<string>) {
+    this.addDatasetCollection = await this.girderResources.getConfiguration(
+      await configIdPromise
+    );
+  }
+
+  get addDatasetDialog() {
+    return this.addDatasetCollection !== null;
+  }
+
+  set addDatasetDialog(value: boolean) {
+    if (!value) {
+      this.addDatasetCollection = null;
+    }
+  }
+
+  addedDatasets() {
+    // New datasets have been added to current collection
+    this.refreshItems(true);
+    // Close the dialog
+    this.addDatasetCollection = null;
   }
 
   @Watch("datasetId")
